+++ webroot/panel/account.php
@@ -105,23 +105,27 @@ if ($_SERVER['REQUEST_METHOD'] == "POST") {
                 UnityHTTPD::messageError("Cannot Submit PI Request", "Already a PI");
                 UnityHTTPD::redirect();
             }
-            if ($SQL->requestExists($USER->uid, UnitySQL::REQUEST_BECOME_PI)) {
+            $group = $USER->getNamesakePIGroup();
+            $request = "$USER->uid:$group->gid";
+            if ($SQL->requestExists($request, UnitySQL::REQUEST_CREATE_PI_GROUP)) {
                 UnityHTTPD::messageError("Cannot Submit PI Request", "This request already exists");
                 UnityHTTPD::redirect();
             }
             if ($_POST["tos"] != "agree") {
                 UnityHTTPD::badRequest("user did not agree to terms of service");
             }
-            $USER->getPIGroup()->requestGroup();
+            $group->requestGroup($USER->uid);
             UnityHTTPD::messageSuccess("PI Group Requested", "");
             UnityHTTPD::redirect();
             break; /** @phpstan-ignore deadCode.unreachable */
         case "cancel_pi_request":
-            if (!$SQL->requestExists($USER->uid, UnitySQL::REQUEST_BECOME_PI)) {
+            $group = $USER->getNamesakePIGroup();
+            $request = "$USER->uid:$group->gid";
+            if (!$SQL->requestExists($request, UnitySQL::REQUEST_CREATE_PI_GROUP)) {
                 UnityHTTPD::messageError("Cannot Cancel PI Request", "No PI request found");
                 UnityHTTPD::redirect();
             }
-            $USER->getPIGroup()->cancelGroupRequest();
+            $group->cancelGroupRequest($USER->uid);
             UnityHTTPD::messageSuccess("PI Request Cancelled", "");
             UnityHTTPD::redirect();
             break; /** @phpstan-ignore deadCode.unreachable */
@@ -178,7 +182,7 @@ if ($isPI) {
         <p>You are currently a <strong>principal investigator</strong> on the Unity HPC Platform.</p>
     ";
 } else {
-    if ($USER->getPIGroup()->exists() && $USER->getPIGroup()->getIsDisabled()) {
+    if ($USER->getNamesakePIGroup()->exists() && $USER->getNamesakePIGroup()->getIsDisabled()) {
         echo "<p>You are no longer a PI because your PI group is disabled.</p>";
     }
     if ($USER->getFlag(UserFlag::QUALIFIED)) {
@@ -217,7 +221,7 @@ if (!$isPI) {
         >
     ";
     echo $CSRFTokenHiddenFormInput;
-    if ($SQL->requestExists($USER->uid, UnitySQL::REQUEST_BECOME_PI)) {
+    if ($SQL->requestExists($USER->uid, UnitySQL::REQUEST_CREATE_PI_GROUP)) {
         $onclick = "return confirm(\"Are you sure you want to cancel this request?\")";
         echo "<input type='submit' value='Cancel PI Account Request' onclick='$onclick'/>";
         echo "
@@ -227,7 +231,7 @@ if (!$isPI) {
             <input type='hidden' name='form_type' value='cancel_pi_request'/>
         ";
     } else {
-        if ($USER->getPIGroup()->exists() && $USER->getPIGroup()->getIsDisabled()) {
+        if ($USER->getNamesakePIGroup()->exists() && $USER->getNamesakePIGroup()->getIsDisabled()) {
             $button_msg = "Request to Re-Enable PI Group";
             $onclick = "return confirm(\"Are you sure you want to re-enable your old PI group?\")";
         } else {