+++ workers/setup-pi-group-owners.php
@@ -1,0 +1,13 @@
+#!/usr/bin/env php
+<?php
+include __DIR__ . "/init.php";
+
+use UnityWebPortal\lib\UnityGroup;
+use UnityWebPortal\lib\UnityLDAP;
+
+foreach ($LDAP->getPIGroupsAttributes(["cn"], filter: UnityLDAP::INCLUDE_DISABLED) as $attributes) {
+    $gid = $attributes["cn"][0];
+    $entry = $LDAP->getPIGroupEntry($gid);
+    $entry->setAttribute("ownerUid", UnityGroup::NamesakeGID2OwnerUID($gid));
+}
+